refactor: merge elements and statements
