+++ lib/orchestrator.dart
@@ -1,7 +1,6 @@
 library orchestrator;
 
+export './src/builder.dart';
 export './src/context.dart';
 export './src/element.dart';
 export './src/emitter.dart';
-export './src/spec.dart';
-export './src/statement.dart';
