+++ lib/src/element.dart
@@ -2,14 +2,17 @@ library element;
 
 import 'package:orchestrator/orchestrator.dart';
 
+part './elements/binary.dart';
 part './elements/class.dart';
 part './elements/constructor.dart';
 part './elements/enum.dart';
 part './elements/extension.dart';
 part './elements/field.dart';
+part './elements/literal.dart';
 part './elements/method.dart';
 part './elements/parameter.dart';
 part './elements/reference.dart';
+part './elements/static.dart';
 
 /// {@template element}
 /// Represents an enclosed piece of source code like a class, enum, or field.
@@ -18,7 +21,10 @@ part './elements/reference.dart';
 /// be generated. Every [Element] has at least one [Emitter] that transforms
 /// its configuration into source code.
 /// {@endtemplate}
-abstract class Element extends Spec {
+abstract class Element {
   /// {@macro element}
   const Element();
+
+  /// Returns an instance of `this` with the provided values.
+  Element copyWith();
 }
