+++ lib/src/emitter.dart
@@ -2,16 +2,18 @@ library emitter;
 
 import 'package:orchestrator/orchestrator.dart';
 
+part './emitters/binary.dart';
 part './emitters/class.dart';
 part './emitters/constructor.dart';
+part './emitters/element.dart';
 part './emitters/enum.dart';
 part './emitters/extension.dart';
 part './emitters/field.dart';
+part './emitters/literal.dart';
 part './emitters/method.dart';
 part './emitters/parameter.dart';
 part './emitters/reference.dart';
-part './emitters/spec.dart';
-part './emitters/statement.dart';
+part './emitters/static.dart';
 
 /// {@template emitter}
 /// The [Emitter] is used to transform any object [T] into an output.
