+++ lib/src/emitters/method.dart
@@ -48,7 +48,7 @@ class MethodEmitter extends Emitter<Method> {
     output.write('{');
 
     for (final v in value.body) {
-      StatementEmitter(context).emit(v, output);
+      ElementEmitter(context).emit(v, output);
     }
 
     output.write(' }');
